refactor: move kafka monitor fixtures to its own tests file
This is to declutter the conftest file.
